requirements.txt
python-dotenv==1.0.0
mysql-connector-python==8.1.0
watchdog==3.0.0